Fix: Apply new color only to td elements.
Previously it was possible that the type of element that is the target of the click on the grid is not actually td. This caused an entire row to get colored at the same time. Sometimes a CSS-BackgroundColor could have even been applied to the #pixelCanvas element, which the discardUnit function could not remove so the grid would be colored in a specific color permanently.